# Remove expensive TRACE logging in find