@bep 2016-09-26
>Is there a way to make the trace logging calls a noop instead of removing them all?

Sure.